requirements.txt
asgiref==3.4.1
Django==3.2.5
django-debug-toolbar==2.2.1
django-extensions==3.1.3
gunicorn==20.0.4
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pytz==2021.3
sqlparse==0.4.2


toml~=0.10.2
elasticsearch~=7.15.1
pydantic~=1.8.2